optimize memory issue

+++ create_embeddings.py
@@ -19,7 +19,7 @@ print("Loading embedding model...")
 model = SentenceTransformer("all-MiniLM-L6-v2")
 
 print("Generating embeddings...")
-embeddings = model.encode(texts, show_progress_bar=True)
+embeddings = model.encode(texts, show_progress_bar=True).astype(np.float32)
 
 np.save("embeddings.npy", embeddings)
